+++ src/Discore/WebSocket/ShardFailureReason.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace Discore.WebSocket
 {
     public enum ShardFailureReason
@@ -8,9 +6,6 @@ namespace Discore.WebSocket
         /// Should be reported to the Discore developers if received.
         /// </summary>
         Unknown,
-
-        [Obsolete("This failure reason can no longer happen, any code handling it can be removed.")]
-        IOError,
         /// <summary>
         /// The shard was invalid, given the sharding settings for the application.
         /// </summary>
@@ -23,6 +18,15 @@ namespace Discore.WebSocket
         /// Occurs if only one shard is used, and that shard would have handled too many guilds.
         /// More than one shard is required if this happens.
         /// </summary>
-        ShardingRequired
+        ShardingRequired,
+        /// <summary>
+        /// A specified intent was invalid. You may have incorrectly calculated the bitwse value.
+        /// </summary>
+        InvalidIntents,
+        /// <summary>
+        /// A disallowed Gateway intent was specified. May happen if you specify an intent that
+        /// you have not enabled or are not approved for (i.e. privileged intents).
+        /// </summary>
+        DisallowedIntents
     }
 }